+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Types of changes
 
 ## [unreleased]
 
+### Changed
+
+-   [Add input text for the quick faceted search](https://github.com/Talend/ui-faceted-search/pull/12)
+
 ### Added
 
 -   [allow to hide badges in the list](https://github.com/Talend/ui-faceted-search/pull/11)